Make ManyToOne bean annotation actually take effect
This annotation was missing runtime retention, so when annotations of methods annotated with this were were retrieved at runtime using reflection, this annotation was never found, meaning it was never really doing anything.
I aligned it with the other annotations in this package. It is now also restricted to methods.